refactor : 공통 type 파일 분리
